requirements.txt
pandas==2.1.4
numpy==1.24.3
psycopg2-binary==2.9.9
shapely==2.0.2
pyproj==3.6.1
python-dotenv==1.0.0
flask==3.0.0
folium==0.15.0
requests==2.31.0